pyproject.toml
[tool.poetry]
name = "nonebot_plugin_bam"
version = "0.0.1"
description = "Bilibili activity monitor plugin for nonebot"
authors = ["7sDream <i@7sdre.am>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/7sDream/nonebot_plugin_bam"
repository = "https://github.com/7sDream/nonebot_plugin_bam"
keywords = ["nonebot", "bilibili", "qqbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-alpha.9"
peewee = "^3.14.0"
aiohttp = {extras = ["speedups"], version = "^3.7.2"}
nonebot-plugin-apscheduler = "^0.1.2"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
